update script deployment logic and simplify wagmi hooks on frontend
