--- frontend/src/hooks/useWaitForGameContract.ts
@@ -1,9 +1,0 @@
-import { useWaitForTransactionReceipt } from "wagmi";
-
-export const useWaitForGameContract = (hash: `0x${string}` | undefined) => {
-  const { isSuccess: hasWaited } = useWaitForTransactionReceipt({
-    hash: hash
-  });
-
-  return { hasWaited };
-};
